pytorch/requirements.txt
# pip install numpy
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
Flask==2.3.1
fonttools==4.39.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lap==0.4.0
lit==16.0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
opencv-python==4.7.0.72
packaging==23.1
pafy==0.5.5
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sympy==1.11.1
thop==0.1.1.post2209072238
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.0
youtube-dl==2020.12.2
waitress==2.1.2